refactor: organize recommendation component
