+++ BookRecommendationSystem/myapp/templates/myapp/about.html
@@ -110,8 +110,8 @@
         <div class="member"><img class="member-image" src="{% static 'images/kornelia.jpeg' %}" alt="Kornelia"><br>
             <h3>Kornelia Madela</h3>
         </div>
-        <div class="member"><img class="member-image" src="{% static 'images/pfp.png' %}" alt="Rita"><br>
-            <h3>Rita Bentinho Braunschweig </h3>
+        <div class="member"><img class="member-image" src="{% static 'images/rita.jpg' %}" alt="Rita"><br>
+            <h3>Rita Braunschweig </h3>
         </div>
         <div class="member"><img class="member-image" src="{% static 'images/pfp.png' %}" alt="Filip"><br>
             <h3>Filip Sima</h3>
